map from image file works
